+++ AlertHub.Api/Options/CacheOptions.cs
@@ -1,0 +1,6 @@
+﻿namespace AlertHub.Api.Options;
+
+public class CacheOptions
+{
+    public TimeSpan AlertExpiry { get; set; } = TimeSpan.FromSeconds(30);
+}