+++ AlertHub.React/src/features/alerts/model/AlertLocationDto.ts
@@ -1,0 +1,6 @@
+export type AlertLocationDto = {
+    Id: string, 
+    Title: string, 
+    Description: string, 
+    Timestamp: Date
+}