+++ AlertHub.React/src/features/alerts/services/mappers/mapLocationsToActiveAlerts.ts
@@ -1,0 +1,20 @@
+import { config } from "../../../../config/config";
+import type { ActiveAlertLocation } from "../../model/ActiveAlertLocation";
+import type { AlertLocationDto } from "../../model/AlertLocationDto";
+import { mapLocation } from "./mapLocations";
+import { mapTitleToAlertType } from "./mapTitleToAlertType";
+
+export function mapLocationsToActiveAlerts(dtos: AlertLocationDto[]): ActiveAlertLocation[] {
+    return dtos.map(dto => {
+        const receivedAtDate = new Date(dto.Timestamp);
+
+        return {
+            id: dto.Id,
+            location: mapLocation(dto.Id),
+            message: dto.Title,
+            type: mapTitleToAlertType(dto.Title),
+            recievedAt: receivedAtDate, 
+            expiresAt: new Date(receivedAtDate.getTime() + config.alertsTtlMs)
+        };
+    });
+}